Track system property updates to help diagnose errors
It is possible to use arbitrary objects as keys and values of the system
properties (via untyped System.getProperties() mutable map). Not all
such objects can be stored in the configuration cache. It might be hard
to track down the code that adds such property so this commit adds some
information to the configuration cache report: the name of the broken
property (stringified, as this can be any object) and the location where
the property was set (though not a precise line number).

Issue: #18432